Limit LfuncKallenbach to 500eV
This improves performance quite a bit. Tu=500eV would be a pretty crap fusion design but it is possible to have DLS simulations like this. Ideally we would warn the user if this is crossed. I think we should do that when we overhaul the rates and get rid of all the individual functions.