Update test to use front_sheath
This makes it consistent with the analytical reference. The change is necessary because the default is now not to use the front sheath model and instead keep the target heat flux as a small fraction of the upstream heat flux. This was done because the front sheath model is hacky and needs more work to be validated, while taking a small fraction of qpllu is consistent with the original Lengyel approach.